Add tests and specs

+++ src/scribo/pinit.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import sys
 
 PROJECT_TEMPLATE = os.path.join(os.path.dirname(__file__), "sample")
 
@@ -14,4 +15,7 @@ def create_project_dir(project_name):
     Create project directory by copying PROJECT_TEMPLATE
     as project_name.
     """
+
+    if os.path.exists(project_name):
+        sys.exit(f"Project: {project_name} already exists!")
     shutil.copytree(PROJECT_TEMPLATE, project_name)
